+++ tests/test_concurrent.py
@@ -12,7 +12,7 @@ from pytest import LogCaptureFixture
 from asphalt.core import (
     Context,
     TaskFactory,
-    require_resource,
+    get_resource_nowait,
     start_background_task_factory,
     start_service_task,
 )
@@ -28,8 +28,8 @@ class TestTaskFactory:
         async with Context():
             default_factory = await start_background_task_factory()
             other_factory = await start_background_task_factory("other")
-            assert default_factory is require_resource(TaskFactory)
-            assert other_factory is require_resource(TaskFactory, "other")
+            assert default_factory is get_resource_nowait(TaskFactory)
+            assert other_factory is get_resource_nowait(TaskFactory, "other")
 
     async def test_start(self) -> None:
         async def taskfunc() -> str: